+++ e2e/transfers/e2e-transfer-suite/src/index.ts
@@ -100,8 +100,8 @@ const test_service = async function (this: any) {
 
         //get all blockchains
 
-        // let spec = 'https://pioneers.dev/spec/swagger.json'
-        let spec = 'http://127.0.0.1:9001/spec/swagger.json'
+        let spec = 'https://pioneers.dev/spec/swagger.json'
+        // let spec = 'http://127.0.0.1:9001/spec/swagger.json'
 
 
         let chains = [
@@ -118,11 +118,11 @@ const test_service = async function (this: any) {
             // 'AVAX',
             // 'BSC',
             // 'XRP', //Testing after fixing ledger_index_current
-            'ETH',
+            // 'ETH',
             // 'MAYA',   //Amount is wrong
             // // 'GNO',
             // 'BCH',
-            // 'BTC',
+            'BTC',
         ]
 
         const allByCaip = chains.map(chainStr => {
@@ -217,16 +217,16 @@ const test_service = async function (this: any) {
         // })
         // log.info(tag,'paths:',paths)
 
-        paths.push({
-          note: ' ETH account 2',
-          networks: [ 'eip155:1', 'eip155:*' ],
-          type: 'address',
-          addressNList: [ 2147483692, 2147483708, 2147483648 ],
-          addressNListMaster: [ 2147483692, 2147483708, 2147483648, 2, 0 ],
-          curve: 'secp256k1',
-          showDisplay: false
-        })
-        log.info(tag,'paths:',paths)
+        // paths.push({
+        //   note: ' ETH account 2',
+        //   networks: [ 'eip155:1', 'eip155:*' ],
+        //   type: 'address',
+        //   addressNList: [ 2147483692, 2147483708, 2147483648 ],
+        //   addressNListMaster: [ 2147483692, 2147483708, 2147483648, 2, 0 ],
+        //   curve: 'secp256k1',
+        //   showDisplay: false
+        // })
+        // log.info(tag,'paths:',paths)
 
         // paths.push({
         //     note:"Bitcoin account 0 segwit (p2sh)",