+++ e2e/reports/intergration-report/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/integration-coins
 
+## 1.13.24
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.27
+
 ## 1.13.23
 
 ### Patch Changes