+++ packages/pioneer/pioneer-cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pioneer-platform/pioneer-cli
 
+## 1.0.6
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.27
+
 ## 1.0.5
 
 ### Patch Changes